Add files via upload
Task: Below are the steps:

----> Build a Number guessing game, in which the user selects a range.
----> Let’s say User selected a range, i.e., from A to B, where A and B belong to Integer.
----> Some random integer will be selected by the system and the user has to guess that integer in the minimum number of guesses
Analysis:

Explanation 1: If the User inputs range, say from 1 to 100. And compiler randomly selected 42 as the integer. And now the guessing game started, so the user entered 50 as his/her first guess. The compiler shows "Try Again! You guessed too high". That means the random number that is chosen, namely 42, does not lie within 50 to 100. That is the significance of guessing half of the range. And once more, the user guesses half of 50. Could you tell why? So the half of 50 is 25. The user enters 25 as his/her second guess. This time the compiler will show "Try Again! You guessed too small". That means the integers less than 25 (from 1 to 25) are useless to be guessed. Now the range for user guessing is shorter, i.e., from 25 to 50. Intelligently! The user guessed half of this range, so that, user guessed 37 as his/her third guess.  This time again the compiler shows the output, "Try Again! You guessed too small". Because it is the fourth time for the user, every time he is getting to a closer range. Now, the guessing range for the user is from 37 to 50, for which the user has guessed 43 as his fourth guess. This time, the compiler will show an output "Try Again! You guessed too high". Therefore, the new range of guessing for users would be from 37 to 43 once more for which the user did guess half of this range, that is 40 as his fifth guess. Now this time compiler shows output, "Try Again! You guessed too small" this time leaving the guess even smaller such that from 41 to 43. And now user has guessed 41 as his sixth guess. Which is incorrect and prints output "Try Again! You guessed too small". And at last, User Guessed the correct number which is 42 as his seventh guess.
 
Total Number of Guesses =7


Explanations 2: Suppose the User inputs range, say from 1 to 50. And compiler randomly chose 42 as the integer. And now the guessing game began. So the half of 50 is 25. User enters 25 as his/her First guess. This time compiler will show, "Try Again! You guessed too small". That means the integers less than 25 (from 1 to 25) are useless to be guessed. Now the range for the user to guess is shorter, i.e., from 25 to 50. Intelligent! User guessed half of this range; so that, the user guesses 37 as his/her second guess.  This time again, the compiler shows the output, "Try Again! You guessed too small". For the user the guessing range is getting smaller by each guess. Now, the guessing range for user is from 37 to 50, for which the user guessed 43 as his/her third guess. This time the compiler will show an output "Try Again! You guessed too high". So, the new guessing range for users will be from 37 to 43, again for which the user guessed the half of this range, that is, 40 as his/her fourth guess. This time, the compiler shows output, saying, "Try Again! You guessed too small". Leaves the guess even smaller such that from 41 to 43. And now the user has guessed 41 as his / her fifth guess. Which is wrong and it shows output saying "Try Again! You guessed too small". And lastly, the User Guessed right number which is 42 as his/her sixth guess.

Total Guesses = 6

So, the minimum number of guesses depends upon range. And the compiler must calculate the minimum number of guessing depends upon the range, on its own. For this, we have a formula:-

 Minimum number of guessing = log2(Upper bound – lower bound + 1)

Step-by-step algorithm:

---> upper bound and lower bound of the range is taken as input from user.
---> Compiler will generate a random integer between the range and store in a variable for future references.
---> A while loop will be initialized for repetitive guessing
---> Again if the user guessed a number which is greater than a randomly selected number then the user gets an output "Try Again! You guessed too high
---> Else If the user guessed a number which is smaller than a randomly selected number, the user gets an output "Try Again! You guessed too small"
---> And if the user guessed in a minimum number of guesses, the user gets a "Congratulations! " Output.
Else if the user didn't guess the integer in the minimum number of guesses, he/she will get "Better Luck Next Time!" output.